test:update junit of metadata.
