+++ .github/workflows/codecov.yml
@@ -5,12 +5,17 @@ name: Codecov
 
 on:
   push:
-    branches: [ main ]
+    branches:
+      - main
+      - 2021.0
+      - 2020.0
+      - greenwich
   pull_request:
-    branches: [ main ]
-
-permissions:
-  contents: write
+    branches:
+      - main
+      - 2021.0
+      - 2020.0
+      - greenwich
 
 jobs:
   build:
